refactor: change path models to infrastructure.models
