+++ src/cinema_platform_django/subscription/admin.py
@@ -1,7 +1,9 @@
 from django.contrib import admin
 
-from src.cinema_platform_django.subscription.models import SubscriptionModel, PlanModel
-
+from src.cinema_platform_django.subscription.infrastructure.models import (
+    PlanModel,
+    SubscriptionModel,
+)
 
 admin.site.register(SubscriptionModel)
 admin.site.register(PlanModel)
