+++ src/cinema_platform_django/subscription/infrastructure/models/__init__.py
@@ -1,0 +1,8 @@
+from src.cinema_platform_django.subscription.infrastructure.models.plans import (
+    PlanModel,
+)
+from src.cinema_platform_django.subscription.infrastructure.models.subscription import (
+    SubscriptionModel,
+)
+
+__all__ = ["PlanModel", "SubscriptionModel"]
